Cleaned up codebase with formatting and renaming.
