+++ src/controllers/beerPosts/delete/deleteBeerById.ts
