+++ src/controllers/beerPosts/read/getBeerById.ts
