+++ src/middleware/auth/checkIfBeerCommentOwner.ts
@@ -22,7 +22,7 @@ const checkIfBeerCommentOwner: BeerCommentMiddlewareFn = async (req, res, next) 
     if (!queriedComment) {
       throw new ServerError('Could not find a comment that id.', 404);
     }
-    
+
     const currentUser = req.currentUser as User;
 
     if (queriedComment.postedBy.id !== currentUser.id) {
