+++ src/routes/beerCommentRoutes.ts
@@ -1,14 +1,19 @@
 import { Router } from 'express';
 
-import createNewComment from '../controllers/beer-comments/create/createNewBeerComment';
-import deleteCommentById from '../controllers/beer-comments/delete/deleteBeerCommentById';
-import getAllComments from '../controllers/beer-comments/read/getAllBeerComments';
-import getCommentById from '../controllers/beer-comments/read/getBeerCommentById';
-import editCommentById from '../controllers/beer-comments/update/editBeerCommentById';
+/* Controllers */
+import createNewComment from '../controllers/beerComments/create/createNewBeerComment';
+import deleteCommentById from '../controllers/beerComments/delete/deleteBeerCommentById';
+import editCommentById from '../controllers/beerComments/update/editBeerCommentById';
+import getAllComments from '../controllers/beerComments/read/getAllBeerComments';
+import getCommentById from '../controllers/beerComments/read/getBeerCommentById';
+
+/* Middleware */
 import checkIfBeerCommentOwner from '../middleware/auth/checkIfBeerCommentOwner';
 import checkIfUserIsConfirmed from '../middleware/auth/checkIfUserIsConfirmed';
 import checkTokens from '../middleware/auth/checkTokens';
 import getCurrentUser from '../middleware/auth/getCurrentUser';
+
+/* Utils */
 import notAllowedError from '../util/error/notAllowedError';
 
 const commentRoutes = Router({ mergeParams: true });
