+++ src/routes/beerImageRoutes.ts
@@ -1,13 +1,16 @@
 import { Router } from 'express';
 
+/* Controllers */
 import processImageData from '../controllers/beerImages/create/processImageData';
 
+/* Middleware */
+import checkIfBeerPostOwner from '../middleware/auth/checkIfBeerPostOwner';
+import checkIfUserIsConfirmed from '../middleware/auth/checkIfUserIsConfirmed';
 import checkTokens from '../middleware/auth/checkTokens';
 import getCurrentUser from '../middleware/auth/getCurrentUser';
-import checkIfBeerPostOwner from '../middleware/auth/checkIfBeerPostOwner';
 
+/* Utils */
 import uploadImage from '../util/imageUpload/uploadImage';
-import checkIfUserIsConfirmed from '../middleware/auth/checkIfUserIsConfirmed';
 import notAllowedError from '../util/error/notAllowedError';
 
 /** Route handler for '/api/beers/:beerId/image'. */
