+++ src/types/index.d.ts
@@ -1,12 +1,12 @@
 import { JwtPayload } from 'jsonwebtoken';
-import User from "../database/model/User";
+import User from '../database/model/User';
 
 declare global {
-     namespace Express {
-         interface Request {
-               currentUser?: User
-               newAccessToken?: string
-               decodedAccessToken?: JwtPayload
-         }
-     }
- }
+  namespace Express {
+    interface Request {
+      currentUser?: User;
+      newAccessToken?: string;
+      decodedAccessToken?: JwtPayload;
+    }
+  }
+}
